Introduce real `MetadataSource` implementations
This commit completes metadata sourcing by properly implementing `MetadataSource` classes, and refactors
`ExternalResourceResolver` to use them.

The behavior of resolving is now different when Gradle metadata is found: instead of merging the Ivy or Maven
descriptor with Gradle metadata, we're now trying them in sequence. The default is:

- Gradle metadata, then
- Ivy (or Maven) metadata, then
- Artifact

Which means that if Gradle metadata is found, we're __not__ trying to download the Ivy/POM descriptor.
Order and enabled sources are configurable but not tested yet.