Optimize ServiceRegistry
The service lookups were showing up as a hot spot in many
configuration time scenarios. The heart of the problem was
that OwnServices was doing too much work on a lookup - it
iterated over all providers, even if their raw type didn't
match the requested type. To work around that we had added
a "hasService" method as well as a CachingProvider in the
past. These somewhat mitigated follow-up requests, but only
made the problem worse on the first lookup of a service.

This change moves the analysis to service registering time.
Services are kept in a multimap structure that associates
them with their raw type. Only the providers with the right
raw type are asked on each lookup. After this change, the
"hasService" method and the CachingProvider were no longer
necessary, as the normal lookup was fast enough.

Another problem was that we created lots of garbage when
walking up the parent hierarchy on a "getAll" request. Each
layer would create a new array list to hold intermediate
results. This has been changed and we now pass the collecting
array from the lowest layer up the parent hierarchy.

The code that was handling cycles in service lookups used a
HashSet, which lead to two problems: We couldn't report the
actual cycle (since we lost the order) and we were paying
the price of hashing, which is not worth it for such tiny
collections. It now uses an ArrayDeque and a "contains" call
instead, which is more efficient and provides better error
reporting.

The handling of service dependencies was also using a HashSet,
which again lead to unnecessary hash operations. An ArrayList
works fine here and uses less CPU cycles.

Finally, CompositeStoppable was crazy expensive, because it used
a CopyOnWriteArrayList and added each Closable one by one. It now
uses simple synchronization instead, greatly reducing GC pressure.

In order to be able to use some of those data structures that are
Java 6+, I moved the Java version check earlier in the chain so we
don't try to set up any services before checking whether we run on
a supported Java version.

The whole DefaultServiceRegistry class should be broken up for
easier understanding, but I didn't go down that path yet to make
this change easier to review.