Ensure MutableVersionConstraint does not have null preferredVersion
While the interface for `VersionConstraint` declared that version was
`@Nullable`, in reality we prevented this when constructing an immutable
version constraint. This meant that `MutableVersionConstraint.asImmutable()`
converted null values to empty strings.

This changes the contract so that `VersionConstraint.preferredVersion` is no
longer `@Nullable`, and an empty string is consistently used to define a
'missing' preferred version, for all implementations of `VersionConstraint`.