Allow an Ivy descriptor to be published with empty dependency version
When no version was specified for a dependency, the `ivy-publish` plugin
would attempt to publish an Ivy file with a missing 'rev' attribute. This
caused XML validation to fail for the generated file.

We now publish a `rev=''` attribute when the dependency version is missing,
allowing these projects to be successfully published using 'ivy-publish'