Load docs fonts from assets.gradle.com CDN
This significantly improves time to first meaningful paint
according to Google's Lighthouse because:
 - WOFF fonts are much smaller than TTF
 - Fonts are loaded in parallel to the rest of CSS, instead of
   inline. Not only increasing the time to load but also giving
   a chance for the browser to layout the DOM in parallel

Google Fonts is blocked in China and makes page loads super slow.

This tells browers to preconnect to assets.gradle.com and
loads fonts from assets.gradle.com as optional, meaning that
browsers should fallback quickly if the font requests or the
network are slow.

Signed-off-by: Eric Wendelin <eric@gradle.com>