Do not merge Gradle module metadata with Ivy/POM metadata
Previously, we would always load the Maven or Ivy metadata first,
and then merge the Gradle module metadata into the result. With this
change, this process is slightly detangled, such that each metadata
source is used independently.

Note that we still load the Ivy/POM file even when the .module file
is found. This is temporary: a later change will fix this and update
all of the resolution tests for the changed interaction expectations.