Do not request Ivy/POM file when module metadata is used
The previous change meant that the Ivy/POM was not _used_ when
a module metadata file was found. With this change, the unusued
file is not requested when it won't be needed.

The change is trivial, but it required changes to our test
infrastructure due to the change expected requests.