Introduce `MetadataSource` implementations
This commit extract the metadata-specific behaviour of `ExternalResourceResolver` subtypes
into separate `MetadataSource` implementations, and refactors `ExternalResourceResolver` to use them.

Each registered `MetadataSource` is attempted in sequence. If metadata is found using an earlier
source, the later sources are not attempted. The default sequence is:

- Gradle metadata (if experimentalFeatures enabled), then
- Ivy (or Maven) metadata, then
- Artifact

This is a pure refactor: no change to behaviour results from this change.