Make file normalization cheaper
On case-insensitive file systems we were trying to
normalize the case of files passed to the file resolver.
The exact reasons for this are lost to history, but it
can't have been correctness, since for correctness one
would have to look at canonical files.

This significantly reduces configuration time on case
insensitive file systems, since we no longer hit the file
system, but only work on the path string. This is in line
with `java.nio.Path.normalize`, which we should use instead
once we drop Java 6 support.