Remove canonicalization from DirectoryFileTree
This removes some configuration time overhead and makes
input/output snapshotting consistent with regards to changes
in symlink. We now always ignore the fact that a file/directory
was replaced by a symlink (but the content stayed the same). There
might be some tasks that actually need to make a difference between
a file and a link to a file, but for those we should provide an
opt-in to canonical snapshotting, as it is quite expensive to do it
consistently (i.e. canonicalize every file in the whole tree).