Allow users to configure the metadata sources for a repository
With this change, a user can configure the `metadataSources` for an Ivy or Maven
repository. It is thus possible to enable or disable any of the built-in sources:
 - gradleMetadata()
 - mavenPom() or ivyDescriptor()
 - artifact()

This will allow users to opt-in to searching for artifacts where metadata is missing,
and to opt-out of searching for Gradle metadata when it's known not to exist.