Cache more file patterns
The caching file pattern factory was not passed around everywhere
it should be, resulting in many patterns being created from scratch.

Still there are plenty of places where we don't have the necessary
context to pass the factory and we *always* create the default exclude
specs, which are quite numerous. For this reason, the normal pattern
spec factory now caches the default excludes, which do not need any
sophistocated cache policy, since they are unlikely to ever change.

We should re-evaluate if we can make our String->Spec conversion cheaper.
Also if we can't make it cheaper, we should try to remove the spec
factory from the PatternSet API and instead pass it in as a parameter
when we actually want to walk the tree.

fixup patterns