Allow parallel service registry access
Instead of locking around every service access,
allow multiple threads to look up services at
the same time. This requires almost no coordination,,
as most of the state inside a registry is completely
independent of the rest. Every provider can protected its
own state without locking all the others. The only
coordination needed is when stopping, because we need
to ensure that all pending requests are served first.
This is achieved using a request counter.