Always resolve test publications with experimentFeatures enabled
Now that we can selectively choose which `metadataSource` to resolve
we can always verify the published metadata files with experimental
features enabled. This simplifies the tests.

One test needed to be changed because an optional dependency being
dragged in by spring-core was changing the resolution result when
experimental features were enabled.