Relax property name requirements (#3739)
Previously, property names for file inputs and outputs had to be valid Java
identifiers, and nested properties were allowed to use dots ('.') to separate
the elements of the property's name. This was required so outputs could be
easily represented as file and directory names inside TAR archives.

This change removes this restriction, and instead escapes property names when
archiving the task outputs.