Explicitly declare which tests require external dependencies
Ideally, those dependencies should be replaced with synthethic dependencies that mimic what they are doing.
In the meantime, the relatively low number of tests which do make use of external dependencies have been
marked as such. This should fix an error on CI, where the Artifactory mirror is not as quite tolerant as
what we allow in the component coordinates.