Moved the some `FileCollection` implementations for efficiently handling task file inputs out of the native plugins and behind the `TaskFileVarFactory` service, where they can be reused, at least internally for now.
There are 2 implementations which are intended to be used for task input properties: a `ConfigurableFileCollection` that allows build logic to configure a bunch of input files, such as C/C++ source files, and a `FileCollection` that represents some  task inputs that are calculated from other inputs, such as the headers of a C/C++ compile task. Both implementations do some in-memory caching during task execution to make snapshotting and task action execution faster.

Improved the implementations a little and added some unit test coverage. The implementations could be further improved, reused in a bunch more places and better integrated into input file snapshotting. This change does just enough so they can be used in the native compile tasks need.