Remove target config from Maven-sourced deps
`MavenDependencyDescriptor.getModuleConfigurations()` was confusing,
as it provided the target scope of the dependency, _plus_ the magic
'optional' configuration name for optional dependencies.

This change removes the concept of a target configuration for a
Maven-sourced dependency: instead we use the scope directly when
building the dependencies for the 'optional' configuration.

foo