Delegate metadata-missing version listing to metadata source
When Gradle fails to list versions using either `maven-metadata.xml`
or listing directories with `ivyPatterns`, the fallback behaviour is
to inspect all `artifactPatterns` to attempt to determine the
available versions. This is particular useful for `flatDir` repositories,
where Gradle will list artifacts by pattern and parse out the versions.

This behaviour has been move to the `DefaultArtifactMetadataSource`.

Signed-off-by: Daz DeBoer <daz@gradle.com>