Convert buildSrc/settings.gradle & buildSrc/build.gradle to Kotlin
To convert a first plugin from Groovy to Kotlin, the `kotlin-dsl`
plugin needs to be applied to buildSrc. This plugin is provided by the
Gradle Kotlin DSL, it configures the Project for building plugins
written using the Kotlin DSL.

But, first things first, this commit converts the buildSrc build scripts
to Kotlin.

Signed-off-by: Paul Merlin <paul@gradle.com>