Serialize custom selection reasons
This commit implements custom component selection reason serialization. It tries to minimize the impact
on the size on disk by serializing "known" reasons in priority, then only writes the specific, custom,
reasons once: if the same reason is used multiple times, it will be indexed and the string will be referred
to with a simple integer index.

This allows the custom selection messages to be displayed in `dependencyInsight`, as shown by the additional
test cases.

Signed-off-by: Cedric Champeau <cedric@gradle.com>