Emit the origin build invocation ID and execution time of tasks reusing outputs.
Previously, we emitted the build invocation ID for both up-to-date and from-cache,
but emitted the original execution time only for from-cache.
This is now unified.

Moreover, these values now reflect the true origin when from-cache outputs are reused as part of incremental build.
Previously, the values from the first build to consider the task up-to-date after a from cache was considered as the origin for subsequent executions. Now, the origin information is kept from the from-cache execution.