Reduce volatile reads in service registry
Enums are initialized atomically, so there is no
reason to make them volatile. All writes are protected
by a lock.

The common case for a service instance is that it is already
initialized. By putting the result of the volatile read in
a local variable, we can optimize this case, so we don't do
two volatile reads (for the if-condition and for the return
statement).