Extract idiomatic test-fixtures plugin to buildSrc
Plugin is applied to all "groovy projects", see `groovyProject.gradle`.

The plugin configures the Project as a test fixtures producer if
`src/testFixtures` is a directory.

The plugin configures the Project as a test fixtures consumer according
to the `testFixtures` extension configuration.

No more Groovy Closure with optional parameters as a Project extra
property.

Test-fixtures configuration rely on a Groovy DSL only feature that
allows to pass a collection of dependency notations when declaring
dependencies. The Kotlin DSL lacks this feature, see
https://github.com/gradle/kotlin-dsl/issues/639

A workaround has been put in place in `build-extensions.kt` in order to
move forward until this is properly fixed in the Kotlin DSL.

Signed-off-by: Paul Merlin <paul@gradle.com>