Derive Java Library variants for libraries published with maven
This changes the interpretation of dependencies defined in maven POMs.
It maps "compile" scope dependencies to the "java-api" variant and
"runtime" scope dependencies to the "java-runtime" variant of
a Java library. This allows users to profit from api/implementation
separation in Java libraries, even if the Gradle module metadata was not
used for publishing.

This changes the resolution result for dependencies with "runtime" scope.
Therefore, this new behavior is only active in experimental mode
in Gradle 4.x.

Signed-off-by: Jendrik Johannes <jendrik@gradle.com>