Remove some wrong timing loggings
The time which was captured for up-to-date checks is wrong, since it
did not include actual snapshotting any more.
The time captured how long it took to load/unpack from the cache now
would have included the time it took to snapshot the inputs, too.
These timings are much better captured by the corresponding build
operations, so we remove them for now.

Signed-off-by: Stefan Wolf <wolf@gradle.com>