Rework variant dependency/constraint/attributes rules
This commit changes how the variant rules (dependencies, constraints or attributes) are
handled. Instead of being indexed by variant name, it's now a generic container, which
checks if the specified `VariantMetadata` applies. This will allow us to implement rules
which apply unconditionnaly on every variant, or on variants which have a specific
attribute set.

Signed-off-by: Cedric Champeau <cedric@gradle.com>