Add a 'buildSrc clean' step to all builds
Results of Groovy compilation in buildSrc can differ slightly.
This can break caching of integTest tasks. We use the SanityCheck CI
build, which runs before all other CI builds, to prime the cache. To
make sure all builds use the results from the cache, and not some
UP-TO-DATE results from the agent, we need to clean buildSrc before
running the main build.

Signed-off-by: Jendrik Johannes <jendrik@gradle.com>