Adapt version detection to JEP-322
JEP-322 is a modification to the Java versioning scheme to accommodate
the time-based release model that the JDK is moving to. Starting with
JDK 10-ea build 36, the changes from this version scheme are in play and
the version dection logic now breaks because of the addition of the
java.version.date to the version output. This commit adjusts for
that. Additionally, future LTS releases of the JDK will contain an
appended " LTS" as well, so this commit adjuts for that too.

Signed-off-by: Jason Tedor <jason@tedor.me>