Add support for showing the resolved variant and its attributes in `ResolvedComponentResult`
This commit adds a `ResolvedNamedVariantResult getVariant()` method on `ResolvedComponentResult`,
which gives access to the resolved variant details. In practice, the details are currently limited
to:

- the name of the variant
- the attributes attached to this variant

The attributes are desugared: if the original variant had strongly-typed attributes, what we get
by the result of this call is an attribute container where all attributes are of type `String`.
As a consequence, it should only be used for diagnostics.

The `ResolveTestFixture` has been adjusted so that we can now (optionally) check the selected
variant and its attributes. One test has been adjusted to check the variant.

Signed-off-by: Cedric Champeau <cedric@gradle.com>