Report Checkstyle violations from xml report, regardless of build pass/fail.
Prior to this, Checkstyle plugin would be silent unless the build failed or
ignoreFailures=true. If thresholds maxErrors and maxWarnings were not met,
reported rule violations in the XML or HTML reports were not mentioned in
build output.

With this change, if Checkstyle is configured to produce an XML report (default),
the report is parsed and violations summarised in the Checkstyle task build output.
This occurs regardless of whether the build failed or not.

Fixed #881.

Signed-off-by: Brett Randall <javabrett@gmail.com>