Fix test that cannot serialize `AttributeContainer` using `JsonOutput`
The test was using `JsonOutput`, which may be enough for simple cases, but now that we have
an `AttributeContainer` in the result, it doesn't handle this type properly and fails with
a stack overflow error. Instead, we only serialize the minimal amount of things required
to make the test pass.

Signed-off-by: Cedric Champeau <cedric@gradle.com>