Rename local variable for clarity
Quite a few tests were using `def variantToTest = variantToTest` which was rather confusing.
The reason for doing this was a weird interaction between Spock and our test, which led to
the `variantToTest` variable not found in the `expectGraph`. This commit renames the local
variable for clarity, but it doesn't change the underlying issue.

Signed-off-by: Cedric Champeau <cedric@gradle.com>