Always honour DSL-defined dependency constraints
When a dependency constraint is defined in the DSL the user has
explicitly opted-in to this feature, and should not require the
experimental-features flag in order to have them take effect.
This change ensures that DSL-defined contstraints are always honoured,
and instead uses the experimentalFeatures flag to determine if Maven
optional dependencies should be treated as dependency constraints.

Signed-off-by: Daz DeBoer <daz@gradle.com>