Make client module work for dependencies that declare variants
For client modules, which is an outdated concept, we now ignore
variants completely and always fall back to configuration selection.
This is required to keep client modules working for dependencies
with pom metadata when the advanced pom support is active. This is,
because the advanced pom support derives runtime and api variants
from the scopes defined in the pom.

Signed-off-by: Jendrik Johannes <jendrik@gradle.com>