Improve attribute matching performance
Only look at extra attributes (that were not requested)
if there was no match after compatibility checking and
disambiguating the requested attributes. This avoids
creating the full attribute set in most cases.

As a result the schema no longer needs to provide a full
list of attributes, which removes some complexity.

This change also removes the adaptive sizing of the matching
table, since it is now so small that being adaptive doesn't
improve performance, but actually makes it worse due to the
added CPU cycles for finding the right slot in the array.