Add reason to `DependencyMetadata`
This allows propagating reasons from `Dependency` to `DependencyMetadata`, which in turn can be used
as `ComponentSelectionReason` descriptions. Doing this lets us display the description in the dependency
insight report.

This commit, however, does **not** add support for serializing reasons, or publishing them. This means
that this only works for direct dependencies so far.

Signed-off-by: Cedric Champeau <cedric@gradle.com>