Fix for transitive Ivy artifact exclusions
For efficiency, we collect separate lists of exclusions that apply to
modules and artifacts: these are applied at different phases of
dependency resolution. Previously, we were being too restrictive when
building the list of artifact excludes: we omitted composite exclusions
that may contain nested artifact exclusions.

All exclusions that _may_ exclude and artifact are now collected and
applied to determine the inclusion of a particular artifact.

Fixes #3951