Polish Gradle warning summary
These warnings are not just about deprecations, but will
include other kinds of warnings in the future. Calling them
Gradle warnings also makes it clear that we are not talking
about Java compiler warnings for instance.

Mentioning that this may break is redundant with what the
individual warnings are already saying and might actually
be wrong. E.g. if there is a single deprecation for Gradle
6.0, then saying "this may break in 5.0" is incorrect. It
also sounds scarier than it needs to be.