Identify host fragment as host in file:////host/path resource names (#4074)
Before, also 'host' was treated as part of the path. With the changes
in e13e45a, we do additional path normalization on resource names
when we resolve a relative path. That is, all additional '/'s are
stripped. This also converted '//host' to '/host'. Which is different:
the first is a host (i.e. a network share), and the second is just the
first segment of a path. This fix identifies a host in a path that
starts with '//' and makes this host part of the 'root' in a
ExternalResourceName rather than the path.

This fixes #4002, which is a 4.4 regression. In practice, these changes
only have an effect on Windows.