Improve `ResolveTestFixture` to support multiple reasons
The resolve test fixture only took the last "description" into consideration. This makes it hard to write tests for custom dependency
reasons. Also, the output of the comparison wasn't very developer friendly, because it was showing the internal, stringy, representation
of the dependency graph nodes.

This commit improves everything by not relying on string comparisons anymore to compare the components. Edges are still compared using
strings, but it already makes things much easier to reason about.