testingJvm: Enhance JUnit test status reporting
Improve status reporting for test classes and individual tests by
populating TestCompleteEvent with the TestResult.ResultType. Change unit
tests accordingly.

Test classes with report FAILURE if the runner failed or at least one
test failed, SKIPPED, if the runner never ran, and SUCCESS otherwise.

This work can be used in the future to optimize the forkEvery behavior
of the RestartEveryNTestClassProcessor to not restart the JVM if the
test runner never ran (issue #3851)

Signed-off-by: Ilya Lipnitskiy <ilya.lipnitskiy@gmail.com>