Show variant details in dependency insight report
This commit introduces the visualization of selected variant details in the
dependency insight report. By default, variant details are not displayed,
and it's required to use the `--show-variant-details` flag to enable them.

Depending on the type of dependency, variant details may either refer to
the selected target configuration, or show proper variant attributes. In
both cases, the selected variant display name is shown. When a variant
defines attributes, the attributes of the variant are displayed and compared
to the requested attributes.

The report will show which attributes where matched, which attributes were
compatible, which attributes weren't requested, and eventually which attributes
were found on the variant, but not used during selection.