Use `ArrayDeque` instead of `LinkedHashSet`
This commit replaces the use of a `LinkedHashSet` with an array deque, considering that in
most cases, there will only be a single descriptor. So we can afford the price of checking
with `contains` when adding descriptors. This should also reduce the memory footprint, and
fix the failing performance test.