Make repository implementations more consistent
The `ModuleComponentRepository` that provides ivy 'dynamic resolve'
was using an anonymous class for implementation, which is not
consistent with other implementations. This change makes this a full-
fledged implementation extending `BaseModuleComponentRepository` like
all of the others.

This refactoring is primary aimed to aid understandability, and precedes
documentation of these types.