[GCS] Upgrade Google GCS library to latest version
** Expected Behavior **
Starting with https://github.com/gradle/gradle/pull/2258 Gradle has supported using a GCS bucket as a
maven repository similar to their S3 support.

Internally we've been using this since before the 4.2 release, with
much success.

** Current Behavior **
However, we've noticed a failure rate in about 1 in 100 to 1 in 1,000
builds (note: a given build might make 100s of GCS requests),
where Gradle loses connection to the data stream that GCS is
returning.

The error appears as follows:
```
* What went wrong:
Could not resolve all files for configuration ':CENSORED:compileClasspath'.
> Could not download google-api-services-datastore-protobuf.jar (com.google.apis:google-api-services-datastore-protobuf:v1beta2-rev1-3.0.0)
   > Could not get resource 'gcs://CENSORED/releases/com/google/apis/google-api-services-datastore-protobuf/v1beta2-rev1-3.0.0/google-api-services-datastore-protobuf-v1beta2-rev1-3.0.0.jar'.
      > Connection closed prematurely: bytesRead = 31425, Content-Length = 445521
```

With a stack trace pointing towards:
https://github.com/google/google-http-java-client/blob/dev/google-http-client/src/main/java/com/google/api/client/http/javanet/NetHttpResponse.java#L202

** Context **
As we're running on Google Compute Engine Hardware we reached
out to Google and they recommended upgrading the library to
rev116.

Internally we compiled a version of the Gradle 4.3 release with the
library upgrade and we've seen the error rate disappear entirely
(no failures for this reason in the last 90,000 builds).

The relevant diff for this upgrade would be:
```
diff --git a/gradle/dependencies.gradle b/gradle/dependencies.gradle
index d0396e3a890..3a91bbec38c 100755
--- a/gradle/dependencies.gradle
+++ b/gradle/dependencies.gradle
@@ -139,7 +139,7 @@ libraries.awsS3 = [
 libraries.zinc = 'com.typesafe.zinc:zinc:0.3.15'

 libraries.gcs = [
-        dependencies.create('com.google.apis:google-api-services-storage:v1-rev78-1.22.0') {
+        dependencies.create('com.google.apis:google-api-services-storage:v1-rev116-1.23.0') {
             exclude group: 'commons-logging', module: 'commons-logging'
         },
         "com.fasterxml.jackson.core:jackson-core:${versions.jackson}",
```

** Steps to Reproduce (for bugs) **
Due to the rarity of this bug, I don't have a compelling mechanism
for reproduction other than creating a GCS repo and running a
couple hundred builds each pulling a couple hundred deps

Signed-off-by: Sam <sam@snapchat.com>