Provide in-memory caching for dependency caches
Instead of using a separate `InMemoryCachedModuleComponentRepository`
for in-memory caching of resolution from remote repositories, the
filesystem-backed cache implementations now each have a build-scoped
in-memory cache as well.

This provides more consistent caching behaviour, since any cache-expiry
logic is shared between in-memory and filesystem caches.

For file-backed repositories, we don't use the filesystem-backed cache,
so we continue to use `InMemoryCachedModuleComponentRepository` for
in-memory caching of resolved modules.

Note that this is only a partial fix, as it does not address the
incorrect caching of maven snapshot modules. This is due to the id-based
caching of resolved artifacts, and the fact that different maven
snapshot instances do not have a different unique identifier.