Show that a constraint was used in the dependency insight report
This commit adds a new cause for a component selection reason, `CONSTRAINT`, which is used
whenever a constraint has been used during selection. This lets us recognize, in the resolution
result API, when a constraint has been applied. The dependency insight report makes use of
that information to show a `via constraint` label in that case.

If a dependency constraint also provides a custom reason, then the message displayed in the
report mentions both `via constraint` and the custom reason.