Do not calculate artifacts for pending dependencies
Pending dependencies (constraints or optional dependencies) are
represented as additional incoming edges for a selected node in the
dependency graph. For these edges, we should not calculate artifacts.
If we do so, we might add additional ones.

For example, this should only give us `foo-1.1-shaded.jar`
and not `foo-1.1.jar` in addition:

dependencies {
    conf 'org:foo:1.0:shaded'
    constraints {
        conf 'org:foo:1.1'
    }
}