Preserve Maven snapshot ID through dependency resolution
When resolving a Maven snapshot that has a unique timestamp, Gradle internally
generates a unique component identifier for this snapshot. Previously this
unique snapshot id was discarded during resolution, meaning that artifacts from
different unique snapshots could have the same identifier.

With this change, the `MavenUniqueSnapshotIdentifier` is preserved in
resolution, and is made available via the resolution result.

This change is required to fix #3019 for maven snapshots, to avoid the
in-memory cache of resolved artifacts providing a stale result for
unique snapshot artifacts.