Remove the `show-variant-details` flag
This commit removes the flag to enable variant details display, and
instead displays the variant details in every case. The rationale is
that usage of `dependencyInsight` is rare, and that without knowing
the existence of the flag, it's unlikely to be used. The additional
details also don't add too much overhead to the output, making it
reasonable to show the variant details by default.