Fix convention mapping class loader leak
The reflection cache keeps a set of property names for each class.
This set was provided by calling keySet() on a map which contains
class values. The keyset would thus keep a strong reference back
to its map and all of its values, leading to a classloader leak.

We now do a defensive copy of they keyset to avoid surprises for
callers of this method.