Add failing test for #4261
In-memory caching of module metadata is done on a per-repository
basis: the same repository definition in different projects will
allow sharing of in-memory cached module metadata.

Component metadata rules are defined on a per-project basis. This
means that in the scenario where different component metadata rules
have been defined for different projects, only the rules from the
first project to resolve a particular module from a particular
repository will apply.

This fix for this issue will require further changes beyond the fix
for #3019.