Cache processed metadata in-memory
Processing component metadata rules for resolved module metadata
is expensive, and the fix for #3019 introduced a serious performance
regression. This change adds in-memory caching for the post-processed
metadata on a per-repository basis.

Unfortunately, this re-introduces the bug #4261, where component
metadata rule outputs can leak between projects for a repository.
Since defining different metadata rules in different projects is
unlikely to be commonplace, this is considered an acceptable
interim position. This issue should be addressed with the
introduction of general-purpose caching of component metadata
rules.