Retry when downloaded Maven distro is invalid
It seems that some Maven mirrors return an
empty response instead of a "not found" when
asked for a nonexisting Maven version. This
made the Maven downloader test fail, as it
tried to unpack this empty distribution.

Instead, we now retry with the next mirror
when the distro can't be unpacked, so we
fail with the expected exception stating that
no mirror returned a valid distro.