Fix dependency constraints showing up as 2 different descriptors
Before this fix, whenever a constraint used a custom message, there were 2
separate descriptors in the list of descriptors in the selection reason: one
was `REQUESTED` with the custom message and one `CONSTRAINT`.

After this commit, the custom message is properly attached to the `CONSTRAINT`
descriptor instead. We also make sure that regular constraints do not lead to
an additional "REQUESTED" descriptor being added.