Do not treat the distributions-dependencies project as Java project
The project does not contain any code and produces and empty jar.
This commit turns the project into a non-Java project to avoid
building the jar. Also, our classpath manifest task assumes that
all projects are Java projects that produce a jar. Hence a
special handling of the "distributionsDependencies" project
is needed in the task.