Reset serializer state before each read or write sequence
Before, the reset was handled internally, assuming that the serializer sees
a sequence of writes, followed by a sequence of reads. But once a read was
complete, and that another read was done on the same file, there was a
possibility to be in an inconsistent state.

To avoid this, resetting the state of the serializer is done externally,
when we know we're going to read or write a sequence of components.