Rework how `HttpServerFixture` handles connection
Previously, a connector was assigned to the server, then the server was
started, and we entered a loop to check if connection was successful. If
not, then we tried again.

However, the initial `start()` could fail, because we already assigned
a connector which was started. If for some reason the port is unavailable,
or another error occurs, then we never entered the retry loop.

This commit fixes the problem by starting the server first, *then* assign
and start a connector.