Make sure that conflict resolution on version still kicks in
It's possible for a module to be both subject to capability conflict resolution AND
version conflict resolution. By making a choice, the capability conflict resolution
enforces a result that we might not want. This commit changes its behavior so that
it _removes_ candidates which are eliminated. The version conflict resolver will
make the final choice. In other words, we:

1. look for capabilities
  * if there are conflicting capabilities and a preference is set, eliminate the non preferred candidates
  * if there's no preference, fail resolution
2. continue with version conflict resolution