Use attribute matching for selection whenever target component has variants
Previously, for a dependency declared in a POM or Ivy descriptor file, we
would _not_ use attribute matching to choose a variant from a local Gradle
project. Since there's no way a POM or Ivy file could reference a Gradle
project directly, this could only occur if the dependency was substituted with
a project dependency.

This is a slight breaking change to the (still) incubating dependency substitution
behaviour, but it makes the behaviour more consistent with regular project
dependencies.
- If the target project configurations do not declare attributes, then the behaviour
  is unchanged.
- If the target project configurations have attributes (ie apply the 'java' plugin')
  then attribute matching will be used to choose the appropriate target configuration.
  This means that the 'api' variant will be selected when resolving a 'compileClasspath'
  and the 'implementation' variant will be selected when resolving a 'runtimeClasspath'.