De-duplicate target confs selected by Ivy dependency
For a dependency declared in an Ivy descriptor file, one or more
target configurations will be chosen for the selected component,
based on the 'confMapping' attribute of the dependency.

In some cases, one chosen configuration may be extended by another
chosen configuration, rendering the first configuration redundant.
With this change, these redundant configurations are removed from
the resolution result.

There is a breaking change here, in that the 'resolved configuration'
resolution result will contain fewer configuration nodes. It will
still contain the same modules and artifacts, in the same order.
A similar change was made for Maven modules in 9df4ba294cd39aff3b55bbe402bb7b7870a24bf4.