Make included build build lifecycle state more explicit and avoid data race.
Previously, each included build dropped its lazily created GradleLauncher reference when it completed. However, cross build task references ultimately access the GradleLauncher of the build dependency. If the included build finished all of its tasks (and dropped its launcher) before a cross build task reference involving it was accessed, the build would be implicitly configured _again_ because it had lost all of its state and appeared to be unconfigured.

Instead, there is now an explicit reset() that needs to be invoked on the included build to prepare it to be built again. This is invoked after build a project for the purpose of providing a build plugin.