Enable attribute matching with improved-pom-support
With 'improved-pom-support' enabled, every Maven module provides
variants, which then triggers attribute matching of variants
when the module is selected.

However, dependencies from an 'improved' POM component to a non-POM
component were still reverting to 'legacy' configuration selection
when the target component did not declare any variants. This PR
updates this behaviour: attribute matching is used to choose a target
variant for any dependency _from_ a Maven module with improved POM
support, regardless of the type (or variant-awareness) of the target.

In practise, this will impact 2 cases (when 'improved-pom-support'
is enabled):
- When a Maven module transitively depends on an Ivy module, the
  'default' configuration of the Ivy module will be chosen instead of
  a combination or 'compile','runtime', 'master' and 'default'.
- When a Maven module transitively depends on a Gradle project via
  dependency substitution, and that Gradle project does not have any
  configuration attributes defined, the 'default' configuration will
  be chosen instead of a combination of 'compile', 'runtime', 'master'
  and 'default'.