Remove eager start for included build controllers
This change removes some legacy 'eager-start' functionality
for included build controllers, that is no longer required for
composite builds.

In early iterations of the feature, tasks for included builds were
executed as they were discovered, and in some cases tasks were
discovered late in the execution cycle. This required a
eager-start approach for included build controllers.

Subsequently, composite builds were improved so that the set of tasks
to execute are known prior to commencing execution, and the included
build lifecycle is more carefully managed. Unfortunately, the old
eager-start code was not removed, leading to included builds being
configured and executed multiple times in a single build invocation.

The only case where an included build should be configured multiple
times is when it provides artifacts for the buildscript classpath,
AND it is involved in the composite build execution. This case is
also handled without the eager-start functionality that is
removed in this commit.