Remove import of dependencies.gradle from buildSrc
With #4167, only one entry from dependencies.gradle was still used
in codqQuality.gradle (groovy variant/version). I think it is fine
to not share this for now. It's not the final solution for how
we want to solve the groovy-all vs groovy-xxx use case anyway.

This is a preparation for making dependencies.gradle more kotlin-dsl
friendly in a next step.