Add test case showing limitation of the current implementation of capabilities
The current implementation of "capability discovery" works well if everyone agrees on the version
to use. However, if they do not, but that a module which gave a preference is evicted (by version
conflict, typically), then the constraint is still kept in memory, and we fail. This can be solved
by adding rules to remove the preference (not implemented yet).