Allow worker actions to return a custom result
This will be necessary to transferm annotation
processing information from the compiler worker
back to the main daemon. The Compiler API already
returns a WorkResult, but until now this result
was discarded and a standard result instance
was returned instead.

This change allows worker actions to implement
Callable<WorkResult> and lets the compiler daemons
make use of this. In subsequent changes the compiler
will use a custom subclass that exposes more information.

This change does not affect the public WorkerExecutor API.