Recompile when annotation processor path changes
This fixes a corner case where removing the last
annotation processor can leave stale generated files
around. This is because the next compilatin will be
incremental and only react to input files that actually
changed.

This is a highly unlikely case in practice, as removing
an annotation processor that is still in use would usually
break the code anyway. The more important reason to do this
change is that it will be required for incremental annotation
processing to work correctly.