Rebase performance tests with large dependency graph
Seems like the caching improvements of #4265 effected all these
test scenarios. Most of them positively. Only the non-parallel,
exclude rule merging test shows a slight regression.

This sets the baseline of all these test scenarios to the same baseline
that we set in commit 3114a for the file repo scenario.