Allow publishing capabiities in Gradle metadata
This commit adds support for publishing capabilities in Gradle metadata. Capabilities
are written if the software component implements the `ComponentWithCapabilities`
interface, which is only the case for Java libraries today. When a Java library is
published, we will automatically convert the capabilities declared in the `dependencies`
block into published capabilities.

In the future, it's likely we will want to either filter or allow overwriting the
capabilities which get published.