Add support for mutating capabilities
This commit introduces capability metadata rules. Those rules let a build
author mutate the capabilities of a published component through a rule. This
makes it possible to associate capabilities to modules which are not using
Gradle metadata. There's however a difference with declaring those capabilities
in the `dependencies { capabilities { ... } }` block:

- rules are only for the current project
- the mutated capabilities are never published (like any other rule, it's
meant for fixups)

So the line should be that if the capabilities are meant for publishing and
consumption by other modules, then they should be defined in a `dependencies`
block, otherwise, if they are meant to fix published capabilities, they should
be declared as rules.

Those rules can typically be used to fix the problem when 2 different modules
say a preference on a capability: this is an error case, that needs the user
input. With a rule, the preference declared by a module can be removed.