Fix for broken publication in docs project
When we added configurations to the docs project to allow other
projects to consume the css directory via publications, we broke
the repository publication of this project as it tries to publish
the css directory as an artifact and blows up.

This fix works around the issue by 1) removing the cssElements
artifacts from the archives configuration to prevent it from being
published and 2) removing directory artifacts from the generated
metadata file and only erroring if the artifact is actually being
published.