Only validate artifacts that will be published
The 'uploadArchives' task generates an ivy file referencing _all_
configurations in a project. However, only a subset of these
configurations are actually published: i.e. have the artifacts
published to the server.

Our validation of published artifacts was somewhat inconsistent:
- Only published artifacts were checked for existence
- All artifacts were verified not to be a directory.

This commit changes this validation so that only artifacts
that will be published are verified. This allows a unpublished
configuration to have a directory artifact.