Avoid deadlock between output event renderer and listener manager (#4661)
Previously, the output event listener that forward build op notifications was downstream of the output event renderer, which has its own lock. It was possible to deadlock between this lock and the global listener manager lock.

Now, the build op notification forwarder is a peer of the renderer, so is not notified under its lock.