Fix incorrect estimate cost of fetching metadata
This commit fixes the caching module component repository, which returned a wrong answer
in case of missing modules. When a module was missing from the repository, the answer
was `FAST`, implying that it was cheap to get metadata for the module. However, if there
are more than one repository, this would prevent other repositories from telling if getting
metadata is cheap.

What happened then is that if a module was missing from all repositories, we would answer
that it's fast to get the metadata, and not perform parallel download of metadata.