Make aggregating annotation processors faster
Aggregating processors need to reprocess all classes they are
interested in on any change. Until now this was ensured by passing
all their source files to the compiler for recompilation.

However, the Java compiler can run annotation processing on class
files too. So instead of recompiling the (unchanged) source files
we now tell the compiler which classes to reprocess. This adds a
small new limitation for aggregating processors - their annotations
need to have at least CLASS retention to be visible in the class
files. The benefit is much faster compilation.