Provide ResolvedVersionConstraint when resolving component id
The 'resolved' version constraint is critical to resolving the component id
for a given selector. In order to honour all constraints in the resolution
process, this `ResolvedVersionConstraint` will be composed of more than
just the constraints for a single selector.

With this change, the `ResolvedVersionConstraint` is constructed
prior to resolving the id, rather than being constructed as part of that
process.