Simplify the collection of selection reasons
It is not necessary to resolve all selectors in order to include them
in the graph. For instance, if the we resolve a fixed version `12`, then
there is no reason to also resolve the version range `[10, 12]`.

For this reason, it should not be necessary to resolve a selector in
order to determine the selection reason. With this change, the selection
reason is no longer part of the `ComponentIdResolveResult`, and is
determined directly when adding the selector to the graph.