Consider all 'reject' constraints in conflict resolution
A component can be selected by multiple dependencies, each of which
may bring version constraints. In conflict resolution, we attempt to
find the component with the highest version that also satisfies all
of the constraints for the module.

Previously, we were only considering one of the depenencies for each
component during conflict resolution, which meant we could incorrectly
choose a component version that should be rejected.

This change fixes #4608 by merging the reject constraints for all
incoming dependencies before using them in conflict resolution.