Deprecate FileCollection.asType()
The method was used to add some special conversions for `FileCollection` in Groovy. In an attempt to remove Groovy-specific functionality from public API the method is now deprecated. The type conversions, while not widely used to start with, will mostly keep working because `FileCollection` implements `Iterable<File>` which Groovy know how to convert to other collections.